Adding a python-friendly API to perform processing
